Initial support for booting 64-bit x86 kernel
- Load kernel by bootboot built images as native 64-bit code; most of
  the arch/platform code is stubbed out but it's enough to get to a
  panic in the VM manager trying to activate paging.

- Basic SMP support; the secondary processors are paused and put in an
  idle loop.

- Various bugfixes in user library code to get it to compile correctly
  on amd64; this includes adding some new header files

+++ kernel/platform/pc64/src/timer/pit.h
@@ -29,7 +29,8 @@ class LegacyPIT {
     public:
         static void disable();
 
-        static float configBusyWait(const uint32_t micros);
+        /// Returns the number of picoseconds the actual wait was for
+        static uint64_t configBusyWait(const uint32_t micros);
         static void busyWait();
 };
 }}
